abstract out WorkspaceAclManager
